Add files via upload
Added function that checks if image about to be compared has already been found in Duplicate array, if not it gets added to sorted array. Appears to be working with small sample group so far.

Also removed previous check if duplicate image has already been found, as can do this check if loop that compares and store image in sorted array.